+++ BO360/BO360/BOMovingParticle.cpp
@@ -36,7 +36,7 @@ void BOMovingParticle::Update(double p_deltaTime)
 	// Rotate if ment to be rotated.
 	if (m_rotate)
 	{
-		m_rotation += (p_deltaTime * m_angleIncrement);
+		m_rotation += (float)(p_deltaTime * m_angleIncrement);
 
 		if (m_rotation > 360)
 		{
@@ -45,8 +45,8 @@ void BOMovingParticle::Update(double p_deltaTime)
 	}
 	
 	// Update position.
-	m_position.x = (m_speed * p_deltaTime) * m_direction.x + m_position.x;
-	m_position.y = (m_speed * p_deltaTime) * m_direction.y + m_position.y;
+	m_position.x = (float)(m_speed * p_deltaTime) * m_direction.x + m_position.x;
+	m_position.y = (float)(m_speed * p_deltaTime) * m_direction.y + m_position.y;
 
 	// Kill the particle if it runs out of time.
 	m_timeLeftS -= p_deltaTime;
@@ -58,7 +58,7 @@ void BOMovingParticle::Update(double p_deltaTime)
 
 void BOMovingParticle::Draw(SDL_Texture* p_texture)
 {
-	int4 l_target = int4(m_position.x, m_position.y, m_size.x/2, m_size.y/2);
+	int4 l_target = int4((int)m_position.x, (int)m_position.y, m_size.x / 2, m_size.y / 2);
 	int4 l_source = int4(0, 0, m_size.x, m_size.y);
 
 	BOGraphicInterface::DrawEx(p_texture, l_source, l_target, m_rotation, int2(l_source.z / 2, l_source.w / 2));
